Trying to implement a greedy algorithm for routing
The idea is for the first ambulance, use greedy algorithm to pick one
patient to save and move there. And then change to that patient's
location and find the next patient to save. When it comes to four, which
means the first ambulance saves four patients, it goes back.

After it goes back, send the second ambulance and reset the time but
mark the one being saved.